move ScriptForm from models to forms (#224)
